cli: flat verb-first redesign — drop sdk/builder umbrellas (#0376)
Refactors the entire bitbadges-cli surface to a flat verb-first tree.
The umbrella nouns `sdk` and `builder` are gone. Every former
`bb cli sdk <verb>` and `bb cli builder <verb>` is now `bb cli <verb>`.

Folded duplicates:
  sdk review + builder review + builder verify + builder validate
    → check (with --depth structural | review | full; default full)
  sdk interpret-tx + sdk interpret-collection + builder explain
    → explain (auto-detects tx vs collection from input shape)
  sdk status + builder doctor → doctor

Generalized:
  builder create-with-burner → deploy --burner
  Required `--burner` flag reserves space for future deploy paths
  (`--from <key>` for chain-binary signing, `--api-broadcast` for
  posting a pre-signed tx) without making any of them the silent
  default.

Hoisted out of `builder`:
  builder templates → build
  builder tools list → tools (plural, lists)
  builder tools call → tool (singular, invokes — kubectl-style)
  builder burner → burner
  builder session → session
  builder resources → resources

Hoisted out of `sdk`:
  sdk docs / skills → docs / skills
  sdk address / alias → address / alias
  sdk lookup-token → lookup
  sdk gen-list-id → gen-list-id

File restructure: one file per top-level command under
src/cli/commands/. Old umbrellas (sdk.ts, builder.ts) removed;
templates.ts → build.ts and createWithBurner.ts → deploy.ts via
git mv to preserve history.

index.ts gets a custom Commander help formatter that emits a
sectioned `--help` (Build & Ship / Indexer / Local State / Discovery
/ Address & lookup / Misc) — the tree is flat, the help layout is
grouped.

Tracked in bitbadges-autopilot backlog #0376. Clean break, no
back-compat aliases — beta status earns the rename.
